refactor: make WebviewErrorDrawer cmp independent
